refactor conditions dir structure
