+++ src/conditions/operator.ts
@@ -1,0 +1,20 @@
+export class Operator {
+  static readonly LOGICAL_OPERATORS: ReadonlyArray<string> = ['and', 'or'];
+  static readonly AND = new Operator('and');
+  static readonly OR = new Operator('or');
+
+  public constructor(public readonly operator: string) {
+    if (!Operator.LOGICAL_OPERATORS.includes(operator)) {
+      throw `"${operator}" is not a valid operator`;
+    }
+    this.operator = operator;
+  }
+
+  toObj() {
+    return { operator: this.operator };
+  }
+
+  static fromObj(obj: Record<string, string>) {
+    return new Operator(obj.operator);
+  }
+}
